Use Debian in place of Alpine Linux as the asciigen docker image base.
Alpine Linux uses musl as its C implementation in place of GNU's glibc
implementation. musl's version of getopt had a few undesired behaviors.
To avoid making changes to the project source to accomodate musl, a
slimmed down version of Debian was chosen as the next best replacement
as the base of the docker image.